refactor: move BuildSignHash to llmq/signing
