+++ src/llmq/signing.h
@@ -263,6 +263,9 @@ void IterateNodesRandom(NodesContainer& nodeStates, Continue&& cont, Callback&& 
     }
 }
 
+uint256 BuildSignHash(Consensus::LLMQType llmqType, const uint256& quorumHash, const uint256& id, const uint256& msgHash);
+
+
 } // namespace llmq
 
 #endif // BITCOIN_LLMQ_SIGNING_H
